Code structure update
Separated code into different files. Updates and added some visuals like search bar and page count.
